+++ examples/example_c.c
@@ -11,7 +11,7 @@
 
 int main() {
   double const data[] = {1.2, 3.4, 5.6, 7.8};
-  size_t const shape[] = {sizeof(data) / sizeof(data[0])};
+  uint64_t const shape[] = {sizeof(data) / sizeof(data[0])};
   cnpypp_npy_save("data_from_c.npy", cnpypp_float64, &data, shape, 1, "w",
                   cnpypp_memory_order_fortran);
